+++ CHANGELOG.md
@@ -50,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - REMIX-5155: Update Kit SDK version to 110.1.1 production.
 - REMIX-5605: Updated light rigs for USD 25.11 compatibility and added a migration for legacy project files.
 - Update hdremix to the latest DXVK that has USD 25.11
+- Update Remix target dependencies: hdremix and omni_core_materials to `ext-650ed92-main`
 
 ### Removed